fix(amcharts5): update tooltip helpers
