+++ src/common/composables/amcharts5/pie-chart-helper.ts
@@ -1,25 +1,30 @@
 import type { Root } from '@amcharts/amcharts5';
 import * as am5 from '@amcharts/amcharts5';
 import type { IPieChartSettings } from '@amcharts/amcharts5/.internal/charts/pie/PieChart';
-import * as am5percent from '@amcharts/amcharts5/percent';
 import type { IPieSeriesSettings } from '@amcharts/amcharts5/percent';
+import * as am5percent from '@amcharts/amcharts5/percent';
 
-export const createPieChart = (root: Root, settings?: IPieChartSettings): { chart: am5percent.PieChart } => {
-    const chart = root.container.children.push(am5percent.PieChart.new(root, {
+import type { Currency } from '@/store/modules/display/config';
+import type { CurrencyRates } from '@/store/modules/display/type';
+
+import { currencyMoneyFormatter } from '@/lib/helper/currency-helper';
+
+import { gray } from '@/styles/colors';
+
+export const createPieChart = (root: Root, settings?: IPieChartSettings): am5percent.PieChart => root.container.children.push(
+    am5percent.PieChart.new(root, {
         layout: root.verticalLayout,
         ...settings,
-    }));
-    return { chart };
-};
+    }),
+);
 
-export const createDonutChart = (root: Root, settings?: IPieChartSettings): { chart: am5percent.PieChart } => {
-    const chart = root.container.children.push(am5percent.PieChart.new(root, {
+export const createDonutChart = (root: Root, settings?: IPieChartSettings): am5percent.PieChart => root.container.children.push(
+    am5percent.PieChart.new(root, {
         layout: root.verticalLayout,
         innerRadius: am5.percent(85),
         ...settings,
-    }));
-    return { chart };
-};
+    }),
+);
 
 export const createPieSeries = (root: Root, chart: am5percent.PieChart, settings?: IPieSeriesSettings): am5percent.PieSeries => {
     const series = chart.series.push(am5percent.PieSeries.new(root, {
@@ -28,8 +33,28 @@ export const createPieSeries = (root: Root, chart: am5percent.PieChart, settings
     series.ticks.template.set('visible', false);
     series.labels.template.set('forceHidden', true);
     series.slices.template.setAll({
+        scale: 1,
         toggleKey: 'none',
-        forceInactive: true,
+    });
+    series.slices.template.states.create('hover', {
+        scale: 1,
     });
     return series;
+};
+
+export const setPieTooltipText = (series: am5percent.PieSeries, tooltip: am5.Tooltip, currency?: Currency, currencyRate?: CurrencyRates): void => {
+    tooltip.label.setAll({
+        fill: am5.color(gray[900]),
+        fontSize: 14,
+    });
+
+    const categoryFieldName = series.get('categoryField') || '';
+    const valueFieldName = series.get('valueField') || '';
+
+    series.slices.template.adapters.add('tooltipText', (_, target) => {
+        let value = target.dataItem?.dataContext?.[valueFieldName];
+        if (currency) value = currencyMoneyFormatter(value, currency, currencyRate);
+        const colorHex = target.get('stroke')?.toString();
+        return `[${colorHex}; fontSize: 10px]●[/] {${categoryFieldName}}: [bold]${value}[/] ({valuePercentTotal.formatNumber("0.00")}%)`;
+    });
 };
